requirements.txt
python-dotenv
tabulate
python-gitlab